Fix fanotify_supported()

+++ src/common.h
@@ -45,6 +45,6 @@ void construct_path_list(int argc,
 
 void warn_inotify_init_error(int fanotify);
 
-bool is_timeout_option_valid(unsigned int* timeout, char* o);
+bool is_timeout_option_valid(long* timeout, char* o);
 
 #endif
